complete overhaul of network_extraction module
