+++ porespy/io/__funcs__.py
@@ -1,8 +1,34 @@
+import pickle
 import numpy as np
 from scipy import ndimage as spim
 from porespy.io.evtk import hl as bp
 import scipy.ndimage as nd
 import skimage.io as io
+from pathlib import Path
+
+
+def to_openpnm(net, filename):
+    r"""
+    Save the result of the `extract_pore_network` function to a file that is
+    suitable for opening in OpenPNM.
+
+    Parameters
+    ----------
+    net : dict
+        The dictionary object produced by `extract_pore_network`
+
+    filename : string or path object
+        The name and location to save the file, which will have `.net` file
+        extension.
+
+    """
+    p = Path(filename)
+    p = p.resolve()
+    # If extension not part of filename
+    if p.suffix == '':
+        p = p.with_suffix('.net')
+    with open(p, 'wb') as f:
+        pickle.dump(net, f)
 
 
 def to_vtk(im, path='./voxvtk', divide=False, downsample=False, voxel_size=1,
