+++ porespy/io/__init__.py
@@ -1,3 +1,4 @@
 from . import evtk
 from .__funcs__ import to_vtk
 from .__funcs__ import to_palabos
+from .__funcs__ import to_openpnm
