+++ porespy/network_extraction/__init__.py
@@ -1,23 +1,14 @@
 r"""
 This module contains functions for extracting pore networks from images.
 
-.. autofunction:: partition_pore_space
-.. autofunction:: align_image_with_openpnm
 .. autofunction:: snow
-.. autofunction:: trim_saddle_points
-.. autofunction:: trim_nearby_peaks
-.. autofunction:: reduce_peaks_to_points
-.. autofunction:: find_peaks
-.. autofunction:: extract_pore_network
+.. autofunction:: snow_dual
+.. autofunction:: regions_to_network
+.. autofunction:: add_boundary_regions
 
 """
 
-from .__funcs__ import partition_pore_space
-from .__funcs__ import align_image_with_openpnm
+from .__funcs__ import add_boundary_regions
+from .__getnet__ import regions_to_network
 from .__snow__ import snow
-from .__snow__ import trim_saddle_points
-from .__snow__ import trim_nearby_peaks
-from .__snow__ import reduce_peaks_to_points
-from .__snow__ import find_peaks
-from .__getnet__ import extract_pore_network
-from .__snow_dual__ import snow_dual_network
+from .__snow_dual__ import snow_dual
